Lower enemy damage to half a Heart; show dungeon coords in dev readout
   Link.hearts becomes a float (quarter-point free at 3 Hearts max) with
   Sim.ENEMY_DAMAGE = 0.5f as the single tunable. The HUD gains a half-Heart
   sprite. The top-right dev readout now appends [dungeon sx,sy] so a new
   seed's dungeon is findable on screen.

+++ src/main/java/com/jmgurr/broadsword/model/Link.java
@@ -3,8 +3,8 @@ package com.jmgurr.broadsword.model;
 /** The hero. Position is always a tile coordinate; no physics. */
 public class Link {
 
-    /** Current health, 0..MAX_HEARTS; 0 means death. */
-    public int hearts = World.MAX_HEARTS;
+    /** Current health in half-Heart steps, 0..MAX_HEARTS; 0 means death. */
+    public float hearts = World.MAX_HEARTS;
 
     public enum Dir {
         // y-down data: +tx = right, +ty = down the screen
